feat: support `dag-json` and `dag-cbor` preview rendering (#68)
It's a source of confusion that the Storacha gateway redirects to
dweb.link for `dag-cbor` and `dag-json` content. This is particularly
highlighted when IPNI is down and dweb.link is not able to pull data
from our gateway - it appears as though customer data is unavailable.
However this is not the case - the data is available over bitswap and
via specifying `?format=raw` in the querystring but freeway simply does
not know how to render a preview HTML page for the data so it responds
with 501, causing the redirect behaviour.

This PR adds a preview HTML page for `dag-cbor` and `dag-json` that is
very similar to the HTML page displayed by ipfs.io and dweb.link for
these kinds of data.

<img width="1059" height="1006" alt="Screenshot 2025-09-10 at 15 20 42"
src="https://github.com/user-attachments/assets/37df783e-69cd-45b8-8028-3bd442c151a7"
/>

I have also updated a few URLs from ipfs.io to ipfs.tech and added an
optional parameter to the handlers allowing the `gatewayDomain` to be
set (which ensures resources can be fetched from the same domain as the
page is served from).

resolves https://github.com/storacha/gateway-lib/issues/28